+++ oshi-core/src/main/java/oshi/software/os/windows/WindowsOperatingSystem.java
@@ -30,6 +30,7 @@ import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.sun.jna.Memory;
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
 import com.sun.jna.platform.win32.Advapi32;
@@ -39,16 +40,21 @@ import com.sun.jna.platform.win32.Kernel32Util;
 import com.sun.jna.platform.win32.Psapi;
 import com.sun.jna.platform.win32.Psapi.PERFORMANCE_INFORMATION;
 import com.sun.jna.platform.win32.Tlhelp32;
-import com.sun.jna.platform.win32.WinBase.FILETIME;
 import com.sun.jna.platform.win32.WinDef.DWORD;
+import com.sun.jna.platform.win32.WinError;
 import com.sun.jna.platform.win32.WinNT;
 import com.sun.jna.platform.win32.WinNT.HANDLE;
 import com.sun.jna.platform.win32.WinNT.HANDLEByReference;
+import com.sun.jna.platform.win32.WinReg;
 import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.PointerByReference;
 
 import oshi.jna.platform.windows.Kernel32;
-import oshi.jna.platform.windows.Kernel32.IO_COUNTERS;
+import oshi.jna.platform.windows.WinPerf.PERF_COUNTER_BLOCK;
+import oshi.jna.platform.windows.WinPerf.PERF_COUNTER_DEFINITION;
+import oshi.jna.platform.windows.WinPerf.PERF_DATA_BLOCK;
+import oshi.jna.platform.windows.WinPerf.PERF_INSTANCE_DEFINITION;
+import oshi.jna.platform.windows.WinPerf.PERF_OBJECT_TYPE;
 import oshi.jna.platform.windows.Wtsapi32;
 import oshi.jna.platform.windows.Wtsapi32.WTS_PROCESS_INFO_EX;
 import oshi.software.common.AbstractOperatingSystem;
@@ -61,27 +67,31 @@ import oshi.util.platform.windows.WmiUtil;
 import oshi.util.platform.windows.WmiUtil.ValueType;
 
 public class WindowsOperatingSystem extends AbstractOperatingSystem {
-
     private static final long serialVersionUID = 1L;
 
     private static final Logger LOG = LoggerFactory.getLogger(WindowsOperatingSystem.class);
 
     // For WMI Process queries for command line
-    private static final String processProperties = "ProcessID,CommandLine";
-    private static final ValueType[] processPropertyTypes = { ValueType.UINT32, ValueType.STRING };
+    private static final String PROCESS_PROPERTIES = "ProcessID,CommandLine";
+    private static final ValueType[] PROCESS_PROPERTY_TYPES = { ValueType.UINT32, ValueType.STRING };
 
-    // For WMI Process queries for private working set
-    private static final String workingSetPrivateProperties = "IDProcess,WorkingSetPrivate";
-    private static final ValueType[] workingSetPrivatePropertyTypes = { ValueType.UINT32, ValueType.STRING };
-
-    // For WMI Process queries for start time, IO counters
-    private static final String perfRawDataProperties = "IDProcess,ElapsedTime,Timestamp_Sys100NS,IOReadBytesPerSec,IOWriteBytesPerSec";
-    private static final ValueType[] perfRawDataPropertyTypes = { ValueType.UINT32, ValueType.STRING, ValueType.STRING,
-            ValueType.STRING, ValueType.STRING };
     /*
-     * LastError
+     * Registry variables to persist
      */
-    private static final int ERROR_ACCESS_DENIED = 5;
+    private int perfDataBufferSize = 8192;
+    private static final int PROCESSOR_INDEX = 230;
+    private static final String PROCESSOR_INDEX_STR = Integer.toString(PROCESSOR_INDEX);
+
+    /*
+     * Registry counter block offsets
+     */
+    private int priorityBaseOffset; // 92
+    private int elapsedTimeOffset; // 96
+    private int idProcessOffset; // 104
+    private int creatingProcessIdOffset; // 108
+    private int ioReadOffset; // 160
+    private int ioWriteOffset; // 168
+    private int workingSetPrivateOffset; // 192
 
     static {
         enableDebugPrivilege();
@@ -96,6 +106,121 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
         this.manufacturer = "Microsoft";
         this.family = "Windows";
         this.version = new WindowsOSVersionInfoEx();
+        initRegistry();
+    }
+
+    private void initRegistry() {
+        // Get the title indices from HKEY_PERFORMANCE_TEXT
+        int priorityBaseIndex = 0;
+        int elapsedTimeIndex = 0;
+        int idProcessIndex = 0;
+        int creatingProcessIdIndex = 0;
+        int ioReadIndex = 0;
+        int ioWriteIndex = 0;
+        int workingSetPrivateIndex = 0;
+        // Query with null pointer to get length
+        IntByReference lpcbData = new IntByReference(0);
+        Pointer pPerfText = null;
+        Advapi32.INSTANCE.RegQueryValueEx(WinReg.HKEY_PERFORMANCE_TEXT, "Counter", 0, null, pPerfText, lpcbData);
+        // Now initialize buffer with length and call again
+        pPerfText = new Memory(lpcbData.getValue());
+        Advapi32.INSTANCE.RegQueryValueEx(WinReg.HKEY_PERFORMANCE_TEXT, "Counter", 0, null, pPerfText, lpcbData);
+        // Text contans a sequence of string pairs, counter # and name
+        int offset = 0;
+        while ((priorityBaseIndex == 0 || elapsedTimeIndex == 0 || idProcessIndex == 0 || creatingProcessIdIndex == 0
+                || ioReadIndex == 0 || ioWriteIndex == 0 || workingSetPrivateIndex == 0)
+                && offset < lpcbData.getValue()) {
+            String index = pPerfText.getWideString(offset);
+            offset += 2 * index.length() + 2;
+            String title = pPerfText.getWideString(offset);
+            offset += 2 * title.length() + 2;
+            switch (title) {
+            case "Priority Base":
+                priorityBaseIndex = ParseUtil.parseIntOrDefault(index, 0);
+                break;
+            case "Elapsed Time":
+                elapsedTimeIndex = ParseUtil.parseIntOrDefault(index, 0);
+                break;
+            case "ID Process":
+                idProcessIndex = ParseUtil.parseIntOrDefault(index, 0);
+                break;
+            case "Creating Process ID":
+                creatingProcessIdIndex = ParseUtil.parseIntOrDefault(index, 0);
+                break;
+            case "IO Read Bytes/sec":
+                ioReadIndex = ParseUtil.parseIntOrDefault(index, 0);
+                break;
+            case "IO Write Bytes/sec":
+                ioWriteIndex = ParseUtil.parseIntOrDefault(index, 0);
+                break;
+            case "Working Set - Private":
+                workingSetPrivateIndex = ParseUtil.parseIntOrDefault(index, 0);
+                break;
+            default:
+                // Ignore other strings
+                break;
+            }
+        }
+        // now load the Process registry to match up the offsets
+        // Sequentially increase the buffer until everything fits.
+        // Save this buffer size for later use
+        lpcbData = new IntByReference(this.perfDataBufferSize);
+        Pointer pPerfData = new Memory(this.perfDataBufferSize);
+        while (WinError.ERROR_MORE_DATA == Advapi32.INSTANCE.RegQueryValueEx(WinReg.HKEY_PERFORMANCE_DATA,
+                PROCESSOR_INDEX_STR, 0, null, pPerfData, lpcbData)) {
+            this.perfDataBufferSize += 4096;
+            lpcbData.setValue(this.perfDataBufferSize);
+            pPerfData = new Memory(this.perfDataBufferSize);
+        }
+
+        PERF_DATA_BLOCK perfData = new PERF_DATA_BLOCK(pPerfData.share(0));
+
+        // See format at
+        // https://msdn.microsoft.com/en-us/library/windows/desktop/aa373105(v=vs.85).aspx
+        // [ ] Object Type
+        // [ ][ ][ ] Multiple counter definitions
+        // Then multiple:
+        // [ ] Instance Definition
+        // [ ] Instance name
+        // [ ] Counter Block
+        // [ ][ ][ ] Counter data for each definition above
+
+        long perfObjectOffset = perfData.HeaderLength;
+
+        // Iterate object types. For Process should only be one here
+        for (int obj = 0; obj < perfData.NumObjectTypes; obj++) {
+            PERF_OBJECT_TYPE perfObject = new PERF_OBJECT_TYPE(pPerfData.share(perfObjectOffset));
+            // Identify where counter definitions start
+            long perfCounterOffset = perfObjectOffset + perfObject.HeaderLength;
+            // If this isn't the Process object, ignore
+            if (perfObject.ObjectNameTitleIndex == PROCESSOR_INDEX) {
+                for (int counter = 0; counter < perfObject.NumCounters; counter++) {
+                    PERF_COUNTER_DEFINITION perfCounter = new PERF_COUNTER_DEFINITION(
+                            pPerfData.share(perfCounterOffset));
+                    if (perfCounter.CounterNameTitleIndex == priorityBaseIndex) {
+                        this.priorityBaseOffset = perfCounter.CounterOffset;
+                    } else if (perfCounter.CounterNameTitleIndex == elapsedTimeIndex) {
+                        this.elapsedTimeOffset = perfCounter.CounterOffset;
+                    } else if (perfCounter.CounterNameTitleIndex == creatingProcessIdIndex) {
+                        this.creatingProcessIdOffset = perfCounter.CounterOffset;
+                    } else if (perfCounter.CounterNameTitleIndex == idProcessIndex) {
+                        this.idProcessOffset = perfCounter.CounterOffset;
+                    } else if (perfCounter.CounterNameTitleIndex == ioReadIndex) {
+                        this.ioReadOffset = perfCounter.CounterOffset;
+                    } else if (perfCounter.CounterNameTitleIndex == ioWriteIndex) {
+                        this.ioWriteOffset = perfCounter.CounterOffset;
+                    } else if (perfCounter.CounterNameTitleIndex == workingSetPrivateIndex) {
+                        this.workingSetPrivateOffset = perfCounter.CounterOffset;
+                    }
+                    // Increment for next Counter
+                    perfCounterOffset += perfCounter.ByteLength;
+                }
+                // We're done, break the loop
+                break;
+            }
+            // Increment for next object (should never need this)
+            perfObjectOffset += perfObject.TotalByteLength;
+        }
     }
 
     /**
@@ -160,7 +285,7 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
 
     /**
      * Private method to do the heavy lifting for all the getProcess functions.
-     * 
+     *
      * @param pids
      *            A collection of pids to query. If null, the entire process
      *            list will be queried.
@@ -170,51 +295,14 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
         long now = System.currentTimeMillis();
         List<OSProcess> procList = new ArrayList<>();
         // define here to avoid object creation overhead later
-        // For process times
-        FILETIME lpCreationTime = new FILETIME();
-        FILETIME lpExitTime = new FILETIME();
-        FILETIME lpKernelTime = new FILETIME();
-        FILETIME lpUserTime = new FILETIME();
-        // For read/write bytes
-        IO_COUNTERS lpIoCounters = new IO_COUNTERS();
         // For groups
         List<String> groupList = new ArrayList<>();
         List<String> groupIDList = new ArrayList<>();
-        // Track when data fails and we need fallback
-        Set<Integer> wmiDataNeeded = new HashSet<>();
 
         int myPid = getProcessId();
         Map<Integer, OSProcess> tempProcessMap = new HashMap<>();
 
-        // Populate WMI process list with Private Working Set size
-        Map<String, List<Object>> procs = null;
-        // If there's a list of pids, filter the WHERE clause
-        if (pids != null) {
-            StringBuilder query = new StringBuilder("WHERE ");
-            for (Integer pid : pids) {
-                query.append(String.format("IDProcess=%d OR ", pid));
-            }
-            query.setLength(query.length() - 3);
-            procs = WmiUtil.selectObjectsFrom(null, "Win32_PerfRawData_PerfProc_Process", workingSetPrivateProperties,
-                    query.toString(), workingSetPrivatePropertyTypes);
-        } else {
-            procs = WmiUtil.selectObjectsFrom(null, "Win32_PerfRawData_PerfProc_Process", workingSetPrivateProperties,
-                    null, workingSetPrivatePropertyTypes);
-            // If querying the whole process list, Remove any non-existent
-            // processes from the global process map/cache
-            List<Object> processIDs = procs.get("IDProcess");
-            List<Integer> pidsToKeep = new ArrayList<>(processIDs.size());
-            for (Object pid : processIDs) {
-                pidsToKeep.add(((Long) pid).intValue());
-            }
-            for (Integer pid : processMap.keySet()) {
-                if (!pidsToKeep.contains(pid)) {
-                    processMap.remove(pid);
-                }
-            }
-        }
-
-        // Get processes from WTS, for additional info
+        // Get processes from WTS
         final PointerByReference ppProcessInfo = new PointerByReference();
         IntByReference pCount = new IntByReference(0);
         if (!Wtsapi32.INSTANCE.WTSEnumerateProcessesEx(Wtsapi32.WTS_CURRENT_SERVER_HANDLE,
@@ -222,10 +310,9 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
             LOG.error("Failed to enumerate Processes. Error code: {}", Kernel32.INSTANCE.GetLastError());
             return procList;
         }
-        // extract the pointed-to pointer
+        // extract the pointed-to pointer and create array
         final Pointer pProcessInfo = ppProcessInfo.getValue();
         final WTS_PROCESS_INFO_EX processInfoRef = new WTS_PROCESS_INFO_EX(pProcessInfo);
-        processInfoRef.read();
         final WTS_PROCESS_INFO_EX[] processInfo = (WTS_PROCESS_INFO_EX[]) processInfoRef.toArray(pCount.getValue());
 
         for (WTS_PROCESS_INFO_EX procInfo : processInfo) {
@@ -249,16 +336,6 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
             final HANDLE pHandle = Kernel32.INSTANCE.OpenProcess(WinNT.PROCESS_QUERY_INFORMATION, false,
                     proc.getProcessID());
             if (pHandle != null) {
-                if (Kernel32.INSTANCE.GetProcessTimes(pHandle, lpCreationTime, lpExitTime, lpKernelTime, lpUserTime)) {
-                    proc.setStartTime(lpCreationTime.toTime());
-                    proc.setUpTime(now - proc.getStartTime());
-                    // Kernel and User time units are already saved
-                }
-                // Read/Write bytes
-                if (Kernel32.INSTANCE.GetProcessIoCounters(pHandle, lpIoCounters)) {
-                    proc.setBytesRead(lpIoCounters.ReadTransferCount);
-                    proc.setBytesWritten(lpIoCounters.WriteTransferCount);
-                }
                 // Full path
                 proc.setPath(Kernel32Util.QueryFullProcessImageName(pHandle, 0));
 
@@ -284,27 +361,21 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
                 } else {
                     int error = Kernel32.INSTANCE.GetLastError();
                     // Access denied errors are common. Fail silently.
-                    if (error != ERROR_ACCESS_DENIED) {
+                    if (error != WinError.ERROR_ACCESS_DENIED) {
                         LOG.error("Failed to get process token for process {}: {}", proc.getProcessID(),
                                 Kernel32.INSTANCE.GetLastError());
                     }
                 }
-            } else {
-                // Get Data from WMI
-                wmiDataNeeded.add(proc.getProcessID());
             }
             Kernel32.INSTANCE.CloseHandle(pHandle);
 
             // TODO: There is no easy way to get ExecutuionState for a process.
             // The WMI value is null. It's possible to get thread Execution
             // State and possibly roll up.
-            proc.setState(OSProcess.State.OTHER);
+            proc.setState(OSProcess.State.RUNNING);
 
             proc.setThreadCount(procInfo.NumberOfThreads.intValue());
             proc.setVirtualSize(procInfo.PagefileUsage.longValue());
-            // This is the Working Set, not the Private Working Set
-            // Set here as a default and overwrite later with data from WMI
-            proc.setResidentSetSize(procInfo.WorkingSetSize.longValue());
 
             proc.setOpenFiles(procInfo.HandleCount.intValue());
             tempProcessMap.put(proc.getProcessID(), proc);
@@ -316,74 +387,13 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
             return procList;
         }
 
-        // Get processes from ToolHelp API for parent PID and priority info
-        Tlhelp32.PROCESSENTRY32.ByReference processEntry = new Tlhelp32.PROCESSENTRY32.ByReference();
-        WinNT.HANDLE snapshot = Kernel32.INSTANCE.CreateToolhelp32Snapshot(Tlhelp32.TH32CS_SNAPPROCESS, new DWORD(0));
-        try {
-            while (Kernel32.INSTANCE.Process32Next(snapshot, processEntry)) {
-                // Only consider processes passed to this method
-                if (!procs.containsKey(processEntry.th32ProcessID.intValue())) {
-                    continue;
-                }
-                if (tempProcessMap.containsKey(processEntry.th32ProcessID.intValue())) {
-                    OSProcess proc = tempProcessMap.get(processEntry.th32ProcessID.intValue());
-                    proc.setThreadCount(processEntry.cntThreads.intValue());
-                    proc.setParentProcessID(processEntry.th32ParentProcessID.intValue());
-                    proc.setPriority(processEntry.pcPriClassBase.intValue());
-                }
-            }
-        } finally {
-            Kernel32.INSTANCE.CloseHandle(snapshot);
-        }
-
-        // Update RSS from Private Working Set value.
-        // Reuse the proc object since we're done with the loop
-        for (int p = 0; p < procs.get("IDProcess").size() - 1; p++) {
-            // Last line "total" appears with PID 0, so iterating to size - 1
-            int pwsPid = ((Long) procs.get("IDProcess").get(p)).intValue();
-            if (tempProcessMap.containsKey(pwsPid)) {
-                OSProcess proc = tempProcessMap.get(pwsPid);
-                proc.setResidentSetSize(
-                        ParseUtil.parseLongOrDefault((String) procs.get("WorkingSetPrivate").get(p), 0L));
-            }
-        }
-
-        // Update start time and IO
-        if (!wmiDataNeeded.isEmpty()) {
-            StringBuilder query = new StringBuilder("WHERE ");
-            for (Integer pid : wmiDataNeeded) {
-                query.append(String.format("IDProcess=%d OR ", pid));
-            }
-            query.setLength(query.length() - 3);
-            Map<String, List<Object>> rawDataProcs = WmiUtil.selectObjectsFrom(null,
-                    "Win32_PerfRawData_PerfProc_Process", perfRawDataProperties, query.toString(),
-                    perfRawDataPropertyTypes);
-            for (int p = 0; p < rawDataProcs.get("IDProcess").size(); p++) {
-                int pid = ((Long) rawDataProcs.get("IDProcess").get(p)).intValue();
-                if (tempProcessMap.containsKey(pid)) {
-                    OSProcess proc = tempProcessMap.get(pid);
-                    proc.setUpTime(
-                            (ParseUtil.parseLongOrDefault((String) rawDataProcs.get("Timestamp_Sys100NS").get(p), 0L)
-                                    - ParseUtil.parseLongOrDefault((String) rawDataProcs.get("ElapsedTime").get(p), 0L))
-                                    / 10000L);
-                    proc.setStartTime(now - proc.getUpTime());
-
-                    proc.setBytesRead(
-                            ParseUtil.parseLongOrDefault((String) rawDataProcs.get("IOReadBytesPerSec").get(p), 0L));
-                    proc.setBytesWritten(
-                            ParseUtil.parseLongOrDefault((String) rawDataProcs.get("IOWriteBytesPerSec").get(p), 0L));
-                    // If start time is newer than cached version, delete cache
-                    if (processMap.containsKey(proc.getProcessID())
-                            && processMap.get(proc.getProcessID()).getStartTime() < proc.getStartTime()) {
-                        processMap.remove(proc.getProcessID());
-                    }
-                }
-            }
-        }
-        // Update Command Line
+        // Get the rest of the data we don't already have from the registry
+        updateRegistryStats(pids, tempProcessMap, now);
+        // Command Line only accessible via WMI.
+        // Utilize cache to only update new processes
         Set<Integer> emptyCommandLines = new HashSet<>();
         for (Integer pid : tempProcessMap.keySet()) {
-            if (processMap.containsKey(pid) && processMap.get(pid).getCommandLine().isEmpty()) {
+            if (this.processMap.containsKey(pid) && this.processMap.get(pid).getCommandLine().isEmpty()) {
                 emptyCommandLines.add(pid);
             }
         }
@@ -394,7 +404,7 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
             }
             query.setLength(query.length() - 3);
             Map<String, List<Object>> commandLineProcs = WmiUtil.selectObjectsFrom(null, "Win32_Process",
-                    processProperties, query.toString(), processPropertyTypes);
+                    PROCESS_PROPERTIES, query.toString(), PROCESS_PROPERTY_TYPES);
             for (int p = 0; p < commandLineProcs.get("ProcessID").size(); p++) {
                 int pid = ((Long) commandLineProcs.get("ProcessID").get(p)).intValue();
                 if (tempProcessMap.containsKey(pid)) {
@@ -405,12 +415,111 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
         }
 
         // Update cached list
+        // TODO: this is ugly as we replace a perfectly good cached value with
+        // something new. Better to grab the cached value earlier in the process
+        // and update it instead of creating new
         for (OSProcess p : tempProcessMap.values()) {
-            processMap.put(p.getProcessID(), p);
+            this.processMap.put(p.getProcessID(), p);
             procList.add(p);
         }
-
         return new ArrayList<>(tempProcessMap.values());
+    }
+
+    private void updateRegistryStats(Collection<Integer> pids, Map<Integer, OSProcess> tempProcessMap, long now) {
+
+        List<Integer> pidsToKeep = new ArrayList<>(tempProcessMap.keySet().size());
+
+        // Sequentially increase the buffer until everything fits.
+        IntByReference lpcbData = new IntByReference(this.perfDataBufferSize);
+        Pointer pPerfData = new Memory(this.perfDataBufferSize);
+        while (WinError.ERROR_MORE_DATA == Advapi32.INSTANCE.RegQueryValueEx(WinReg.HKEY_PERFORMANCE_DATA,
+                PROCESSOR_INDEX_STR, 0, null, pPerfData, lpcbData)) {
+            this.perfDataBufferSize += 4096;
+            lpcbData.setValue(this.perfDataBufferSize);
+            pPerfData = new Memory(this.perfDataBufferSize);
+        }
+
+        PERF_DATA_BLOCK perfData = new PERF_DATA_BLOCK(pPerfData.share(0));
+        long perfTime100nSec = perfData.PerfTime100nSec.getValue();
+
+        // See format at
+        // https://msdn.microsoft.com/en-us/library/windows/desktop/aa373105(v=vs.85).aspx
+        // [ ] Object Type
+        // [ ][ ][ ] Multiple counter definitions
+        // Then multiple:
+        // [ ] Instance Definition
+        // [ ] Instance name
+        // [ ] Counter Block
+        // [ ][ ][ ] Counter data for each definition above
+
+        long perfObjectOffset = perfData.HeaderLength;
+
+        // Iterate object types. For Process should only be one here
+        for (int obj = 0; obj < perfData.NumObjectTypes; obj++) {
+            PERF_OBJECT_TYPE perfObject = new PERF_OBJECT_TYPE(pPerfData.share(perfObjectOffset));
+            // If this isn't the Process object, ignore
+            if (perfObject.ObjectNameTitleIndex == PROCESSOR_INDEX) {
+                // Skip over counter definitions
+                // There will be many of these, this points to the first one
+                long perfInstanceOffset = perfObjectOffset + perfObject.DefinitionLength;
+
+                // We need this for every process, initialize outside loop to
+                // save overhead
+                PERF_COUNTER_BLOCK perfCounterBlock = null;
+                // Iterate instances.
+                // The last instance is _Total so subtract 1 from max
+                for (int inst = 0; inst < perfObject.NumInstances - 1; inst++) {
+                    PERF_INSTANCE_DEFINITION perfInstance = new PERF_INSTANCE_DEFINITION(
+                            pPerfData.share(perfInstanceOffset));
+                    long perfCounterBlockOffset = perfInstanceOffset + perfInstance.ByteLength;
+
+                    int pid = pPerfData.getInt(perfCounterBlockOffset + this.idProcessOffset);
+                    if ((pids == null || pids.contains(pid)) && tempProcessMap.containsKey(pid)) {
+                        pidsToKeep.add(pid);
+                        OSProcess proc = tempProcessMap.get(pid);
+
+                        // If name is empty, populate from registry
+                        if (proc.getName().isEmpty()) {
+                            proc.setName(pPerfData.getWideString(perfInstanceOffset + perfInstance.NameOffset));
+                        }
+
+                        proc.setUpTime(
+                                (perfTime100nSec - pPerfData.getLong(perfCounterBlockOffset + this.elapsedTimeOffset))
+                                        / 10000L);
+                        proc.setStartTime(now - proc.getUpTime());
+                        proc.setBytesRead(pPerfData.getLong(perfCounterBlockOffset + this.ioReadOffset));
+                        proc.setBytesWritten(pPerfData.getLong(perfCounterBlockOffset + this.ioWriteOffset));
+                        proc.setResidentSetSize(
+                                pPerfData.getLong(perfCounterBlockOffset + this.workingSetPrivateOffset));
+                        proc.setParentProcessID(
+                                pPerfData.getInt(perfCounterBlockOffset + this.creatingProcessIdOffset));
+                        proc.setPriority(pPerfData.getInt(perfCounterBlockOffset + this.priorityBaseOffset));
+
+                        // If start time is newer than cached version, delete
+                        // cache
+                        if (this.processMap.containsKey(proc.getProcessID())
+                                && this.processMap.get(proc.getProcessID()).getStartTime() < proc.getStartTime()) {
+                            this.processMap.remove(proc.getProcessID());
+                        }
+                    }
+
+                    // Increment to next instance
+                    perfCounterBlock = new PERF_COUNTER_BLOCK(pPerfData.share(perfCounterBlockOffset));
+                    perfInstanceOffset = perfCounterBlockOffset + perfCounterBlock.ByteLength;
+                }
+                // We're done, break the loop
+                break;
+            }
+            // Increment for next object (should never need this)
+            perfObjectOffset += perfObject.TotalByteLength;
+        }
+        if (pids == null) {
+            for (Integer pid : this.processMap.keySet()) {
+                if (!pidsToKeep.contains(pid)) {
+                    this.processMap.remove(pid);
+                }
+            }
+        }
     }
 
     /**
@@ -481,5 +590,4 @@ public class WindowsOperatingSystem extends AbstractOperatingSystem {
         }
         Kernel32.INSTANCE.CloseHandle(hToken.getValue());
     }
-
 }